+++ ava.config.js
@@ -1,0 +1,8 @@
+export default {
+  verbose: true,
+  timeout: '1m',
+  files: ['test/test-*'],
+  require: [
+    "@babel/register"
+  ]
+}