+++ test/test-cli.js
@@ -6,15 +6,18 @@ import pathExists from 'path-exists';
 import readPkg from 'read-pkg';
 // import copy from 'cpy';
 import tempfile from 'tempfile';
+import rimraf from 'rimraf';
 
 const cli = path.resolve('lib/cli.js');
-const read = file => new Promise((resolve, reject) => fs.readFile(file, 'utf8', (error, data) => {
-  if (error) {
-    return reject(error);
-  }
+const read = file => new Promise((resolve, reject) => {
+  fs.readFile(file, 'utf8', (error, data) => {
+    if (error) {
+      return reject(error);
+    }
 
-  resolve(data);
-}));
+    resolve(data);
+  });
+});
 
 test('Check version', async t => {
   const {stdout} = await execa(cli, ['-v']);
@@ -22,7 +25,7 @@ test('Check version', async t => {
   t.is(stdout, version);
 });
 
-test('Transform html witch config in package.json', async t => {
+test('Transform html with config in package.json', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, ['test/fixtures/input.html', '-o', filename]);
@@ -30,7 +33,7 @@ test('Transform html witch config in package.json', async t => {
   t.is((await read('test/expected/output-config-pkg.html')), (await read(filename)));
 });
 
-test('Transform html witch indent', async t => {
+test('Transform html with indent', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, ['test/fixtures/input-indent.html', '-o', filename]);
@@ -38,7 +41,7 @@ test('Transform html witch indent', async t => {
   t.is((await read('test/expected/output-indent.html')), (await read(filename)));
 });
 
-test('Transform html witch config in file', async t => {
+test('Transform html with config in file', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, ['test/fixtures/input.html', '-o', filename, '-c', 'test/fixtures/config.json']);
@@ -46,7 +49,7 @@ test('Transform html witch config in file', async t => {
   t.is((await read('test/expected/output-config-file.html')), (await read(filename)));
 });
 
-test('Transform html witch dot config in file', async t => {
+test('Transform html with dot config in file', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, ['test/fixtures/input.html', '-o', filename, '-c', 'test/fixtures/.config']);
@@ -62,7 +65,7 @@ test('Transform html from two file', async t => {
   t.is((await read('test/expected/output-indent.html')), (await read(`${folder}/input-indent.html`)));
 });
 
-// test('Transform html witch options replace', async t => {
+// test('Transform html with options replace', async t => {
 //   t.plan(2);
 //   const folder = await tempfile();
 //   await copy(['test/fixtures/input.html', 'test/fixtures/input-indent.html'], folder);
@@ -71,7 +74,7 @@ test('Transform html from two file', async t => {
 //   t.is((await read('test/expected/output-indent.html')), (await read(`${folder}/input-indent.html`)));
 // });
 
-test('Transform html witch config in file and stdin options use', async t => {
+test('Transform html with config in file and stdin options use', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, [
@@ -91,7 +94,7 @@ test('Transform html witch config in file and stdin options use', async t => {
   t.is((await read('test/expected/output-bem.html')), (await read(filename)));
 });
 
-test('Transform html witch stdin options use', async t => {
+test('Transform html with stdin options use', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, [
@@ -107,7 +110,7 @@ test('Transform html witch stdin options use', async t => {
   t.is((await read('test/expected/output-custom-elements.html')), (await read(filename)));
 });
 
-test('Transform html witch stdin options use two key', async t => {
+test('Transform html with stdin options use two key', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, [
@@ -129,7 +132,7 @@ test('Transform html witch stdin options use two key', async t => {
   t.is((await read('test/expected/output-bem.html')), (await read(filename)));
 });
 
-test('Transform html stdin options use witch modules', async t => {
+test('Transform html stdin options use with modules', async t => {
   t.plan(2);
   const filename = tempfile('.html');
   await execa(cli, [
@@ -145,59 +148,176 @@ test('Transform html stdin options use witch modules', async t => {
   t.is((await read('test/expected/output-modules.html')), (await read(filename)));
 });
 
+test('Transform html stdin options', async t => {
+  const outputPath = 'test/expected/by-config/options/output.html';
+  rimraf.sync(outputPath);
+  t.plan(2);
+  await execa(cli, [
+    '-c',
+    'test/fixtures/by-config/options/.config'
+  ]);
+  t.true(await pathExists(outputPath));
+  t.is(
+    (await read(outputPath)),
+    (await read('test/fixtures/by-config/options/input.html'))
+  );
+});
+
 test('Transform html stdin options only config one-io', async t => {
+  const outputPath = 'test/expected/by-config/one-io/output.html';
+  rimraf.sync(outputPath);
   t.plan(2);
   await execa(cli, [
     '-c',
     'test/fixtures/by-config/one-io/.config'
   ]);
-  t.true(await pathExists('test/expected/by-config/one-io/output.html'));
+  t.true(await pathExists(outputPath));
   t.is(
-    (await read('test/expected/by-config/one-io/output.html')),
+    (await read(outputPath)),
     (await read('test/fixtures/by-config/one-io/input.html'))
   );
 });
 
 test('Transform html stdin options only config two-io to dir', async t => {
+  const outputPath1 = 'test/expected/by-config/two-io/input-1.html';
+  const outputPath2 = 'test/expected/by-config/two-io/input-2.html';
+  rimraf.sync(outputPath1);
+  rimraf.sync(outputPath2);
   t.plan(4);
   await execa(cli, [
     '-c',
     'test/fixtures/by-config/two-io/.config'
   ]);
-  t.true(await pathExists('test/expected/by-config/two-io/input-1.html'));
-  t.true(await pathExists('test/expected/by-config/two-io/input-2.html'));
+  t.true(await pathExists(outputPath1));
+  t.true(await pathExists(outputPath2));
   t.is(
-    (await read('test/expected/by-config/two-io/input-1.html')),
+    (await read(outputPath1)),
     (await read('test/fixtures/by-config/two-io/input-1.html'))
   );
   t.is(
-    (await read('test/expected/by-config/two-io/input-2.html')),
+    (await read(outputPath2)),
     (await read('test/fixtures/by-config/two-io/input-2.html'))
   );
 });
 
 test('Transform html stdin options only config one-io-by-pattern', async t => {
+  const outputPath = 'test/expected/by-config/one-io-by-pattern/input-1.html';
+  rimraf.sync(outputPath);
   t.plan(2);
   await execa(cli, [
     '-c',
     'test/fixtures/by-config/one-io-by-pattern/.config'
   ]);
-  t.true(await pathExists('test/expected/by-config/one-io-by-pattern/input-1.html'));
+  t.true(await pathExists(outputPath));
   t.is(
-    (await read('test/expected/by-config/one-io-by-pattern/input-1.html')),
+    (await read(outputPath)),
     (await read('test/fixtures/by-config/one-io-by-pattern/input-1.html'))
   );
 });
 
 test('Transform html stdin options only config one-io anf plugins array', async t => {
+  const outputPath = 'test/expected/by-config/one-io-and-plugins-array/output.html';
+  rimraf.sync(outputPath);
   t.plan(2);
   await execa(cli, [
     '-c',
     'test/fixtures/by-config/one-io-and-plugins-array/posthtml.config.js'
   ]);
-  t.true(await pathExists('test/expected/by-config/one-io-and-plugins-array/output.html'));
+  t.true(await pathExists(outputPath));
   t.is(
-    (await read('test/expected/by-config/one-io-and-plugins-array/output.html')),
+    (await read(outputPath)),
     (await read('test/fixtures/by-config/one-io-and-plugins-array/input.html'))
+  );
+});
+
+test('Output with keeping the folder structure with allInOutput option', async t => {
+  const outputPath = 'test/expected/output-nesting';
+  rimraf.sync(outputPath);
+  t.plan(3);
+  await execa(cli, [
+    'test/fixtures/input-nesting/**/*.html',
+    '-o',
+    outputPath,
+    '-a'
+  ]);
+  t.true(await pathExists(outputPath));
+  t.is(
+    (await read('test/fixtures/input-nesting/input-nesting.html')),
+    (await read(`${outputPath}/test/fixtures/input-nesting/input-nesting.html`))
+  );
+  t.is(
+    (await read('test/fixtures/input-nesting/input-nesting-child/input-nesting.html')),
+    (await read(`${outputPath}/test/fixtures/input-nesting/input-nesting-child/input-nesting.html`))
+  );
+});
+
+test('Specify the root of the output folder structure with root option', async t => {
+  const outputPath = 'test/expected/output-nesting-root';
+  rimraf.sync(outputPath);
+  t.plan(3);
+  await execa(cli, [
+    '**/*.html',
+    '-o',
+    outputPath,
+    '-a',
+    '-r',
+    'test/fixtures/input-nesting'
+  ]);
+  t.true(await pathExists(outputPath));
+  t.is(
+    (await read('test/fixtures/input-nesting/input-nesting.html')),
+    (await read(`${outputPath}/input-nesting.html`))
+  );
+  t.is(
+    (await read('test/fixtures/input-nesting/input-nesting-child/input-nesting.html')),
+    (await read(`${outputPath}/input-nesting-child/input-nesting.html`))
+  );
+});
+
+test('Ignoring files by pattern', async t => {
+  const outputPath = 'test/expected/output-ignoring';
+  rimraf.sync(outputPath);
+  t.plan(3);
+  await execa(cli, [
+    '**/*.html',
+    '!ignoring-input-child/**/*.html',
+    '-o',
+    outputPath,
+    '-a',
+    '-r',
+    'test/fixtures/input-ignoring'
+  ]);
+  t.true(await pathExists(outputPath));
+  t.is(
+    (await read('test/fixtures/input-ignoring/input.html')),
+    (await read(`${outputPath}/input.html`))
+  );
+  t.false(await pathExists('test/expected/input-ignoring/ignoring-input-child'));
+});
+
+test('Skip parsing file', async t => {
+  const outputPath = 'test/expected/output-skip';
+  rimraf.sync(outputPath);
+  t.plan(3);
+  await execa(cli, [
+    '**/*.html',
+    '-o',
+    outputPath,
+    '-a',
+    '-r',
+    'test/fixtures/input-skip',
+    '-s',
+    'input-skip.html',
+    '-u',
+    'posthtml-custom-elements'
+  ]);
+  t.true(await pathExists(outputPath));
+  t.is(
+    (await read('test/fixtures/input-skip/input.html')),
+    (await read(`${outputPath}/input.html`))
+  );
+  t.is(
+    (await read('test/fixtures/input-skip/input-skip.html')),
+    (await read(`${outputPath}/input-skip.html`))
   );
 });